requirements.txt
# Core dependencies
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
streamlit>=1.35.0
